+++ tests/testthat/test-estimate_erf.R
@@ -16,11 +16,11 @@ test_that("estimate_erf works as expected", {
                                delta_n = 1,
                                scale = 0.5)
 
-  min_w <- min(pseudo_pop$w)
-  max_w <- max(pseudo_pop$w)
+  min_w <- min(pseudo_pop$pseudo_pop$w)
+  max_w <- max(pseudo_pop$pseudo_pop$w)
 
-  res <- estimate_erf(pseudo_pop$Y,
-                      pseudo_pop$w,
+  res <- estimate_erf(pseudo_pop$pseudo_pop$Y,
+                      pseudo_pop$pseudo_pop$w,
                       bw_seq=seq(0.2,2,0.2),
                       w_vals=seq(min_w,max_w,0.5))